gles compat: Handle glDrawRangeElements with client pointers.
Fixes: #934
