+++ gapis/api/gles/compat.go
@@ -22,9 +22,7 @@ import (
 
 	"github.com/google/gapid/core/data/deep"
 	"github.com/google/gapid/core/log"
-	"github.com/google/gapid/core/math/interval"
 	"github.com/google/gapid/core/math/u32"
-	"github.com/google/gapid/core/math/u64"
 	"github.com/google/gapid/core/os/device"
 	"github.com/google/gapid/gapis/api"
 	"github.com/google/gapid/gapis/api/gles/glsl/ast"
@@ -504,64 +502,17 @@ func compat(ctx context.Context, device *device.Instance) (transform.Transformer
 					moveClientVBsToVAs(ctx, t, clientVAs, first, count, id, cmd, s, c, out)
 				}
 			}
-			MultiviewDraw(ctx, id, cmd, out)
+			compatMultiviewDraw(ctx, id, cmd, out)
 			return
 
-		case *GlDrawElements:
-			if target.vertexArrayObjects == required {
-				e := externs{ctx: ctx, cmd: cmd, s: s}
+		case drawElements:
+			if target.vertexArrayObjects != required {
+				compatMultiviewDraw(ctx, id, cmd, out)
+			} else {
 				t := newTweaker(out, dID, cb)
 				defer t.revert(ctx)
-
-				ib := c.Bound.VertexArray.ElementArrayBuffer
-				clientIB := ib == nil
-				clientVB := clientVAsBound(c, clientVAs)
-				if clientIB {
-					// The indices for the glDrawElements call is in client memory.
-					// We need to move this into a temporary buffer.
-
-					// Generate a new element array buffer and bind it.
-					bufID := t.glGenBuffer(ctx)
-					t.GlBindBuffer_ElementArrayBuffer(ctx, bufID)
-
-					// By moving the draw call's observations earlier, populate the element array buffer.
-					size, base := DataTypeSize(cmd.IndicesType)*int(cmd.IndicesCount), memory.Pointer(cmd.Indices)
-					glBufferData := cb.GlBufferData(GLenum_GL_ELEMENT_ARRAY_BUFFER, GLsizeiptr(size), memory.Pointer(base), GLenum_GL_STATIC_DRAW)
-					glBufferData.extras = cmd.extras
-					out.MutateAndWrite(ctx, dID, glBufferData)
-
-					if clientVB {
-						// Some of the vertex arrays for the glDrawElements call is in
-						// client memory and we need to move this into temporary buffer(s).
-						// The indices are also in client memory, so we need to apply the
-						// command's reads now so that the indices can be read from the
-						// application pool.
-						cmd.Extras().Observations().ApplyReads(s.Memory.ApplicationPool())
-						indexSize := DataTypeSize(cmd.IndicesType)
-						data := U8ᵖ(cmd.Indices).Slice(0, uint64(indexSize*int(cmd.IndicesCount)), s.MemoryLayout)
-						limits := e.calcIndexLimits(data, indexSize)
-						moveClientVBsToVAs(ctx, t, clientVAs, limits.First, limits.Count, id, cmd, s, c, out)
-					}
-
-					cmd := *cmd
-					cmd.Indices.addr = 0
-					MultiviewDraw(ctx, id, &cmd, out)
-					return
-
-				} else if clientVB { // GL_ELEMENT_ARRAY_BUFFER is bound
-					// Some of the vertex arrays for the glDrawElements call is in
-					// client memory and we need to move this into temporary buffer(s).
-					// The indices are server-side, so can just be read from the internal
-					// pooled buffer.
-					data := ib.Data
-					indexSize := DataTypeSize(cmd.IndicesType)
-					start := u64.Min(cmd.Indices.addr, data.count)                               // Clamp
-					end := u64.Min(start+uint64(indexSize)*uint64(cmd.IndicesCount), data.count) // Clamp
-					limits := e.calcIndexLimits(data.Slice(start, end, s.MemoryLayout), indexSize)
-					moveClientVBsToVAs(ctx, t, clientVAs, limits.First, limits.Count, id, cmd, s, c, out)
-				}
+				compatDrawElements(ctx, t, clientVAs, id, cmd, s, out)
 			}
-			MultiviewDraw(ctx, id, cmd, out)
 			return
 
 		case *GlCompressedTexImage2D:
@@ -1108,14 +1059,14 @@ func compat(ctx context.Context, device *device.Instance) (transform.Transformer
 
 		default:
 			if cmd.CmdFlags().IsClear() {
-				MultiviewDraw(ctx, id, cmd, out)
+				compatMultiviewDraw(ctx, id, cmd, out)
 				return
 			}
 			if cmd.CmdFlags().IsDrawCall() {
 				if clientVAsBound(c, clientVAs) {
 					log.W(ctx, "Draw call with client-pointers not handled by the compatability layer. Command: %v", cmd)
 				}
-				MultiviewDraw(ctx, id, cmd, out)
+				compatMultiviewDraw(ctx, id, cmd, out)
 				return
 			}
 		}
@@ -1127,7 +1078,7 @@ func compat(ctx context.Context, device *device.Instance) (transform.Transformer
 }
 
 // Naive multiview implementation - invoke each draw call several times with different layers
-func MultiviewDraw(ctx context.Context, id api.CmdID, cmd api.Cmd, out transform.Writer) {
+func compatMultiviewDraw(ctx context.Context, id api.CmdID, cmd api.Cmd, out transform.Writer) {
 	s := out.State()
 	c := GetContext(s, cmd.Thread())
 	dID := id.Derived()
@@ -1178,104 +1129,4 @@ func (fb *Framebuffer) ForEachAttachment(action func(GLenum, FramebufferAttachme
 // captured with the context c can be replayed on the device d.
 func canUsePrecompiledShader(c *Context, d *device.OpenGLDriver) bool {
 	return c.Constants.Vendor == d.Vendor && c.Constants.Version == d.Version
-}
-
-// clientVAsBound returns true if there are any vertex attribute arrays enabled
-// with pointers to client-side memory.
-func clientVAsBound(c *Context, clientVAs map[*VertexAttributeArray]*GlVertexAttribPointer) bool {
-	for _, arr := range c.Bound.VertexArray.VertexAttributeArrays {
-		if arr.Enabled == GLboolean_GL_TRUE {
-			if _, ok := clientVAs[arr]; ok {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// moveClientVBsToVAs is a compatability helper for transforming client-side
-// vertex array data (which is not supported by glVertexAttribPointer in later
-// versions of GL), into array-buffers.
-func moveClientVBsToVAs(
-	ctx context.Context,
-	t *tweaker,
-	clientVAs map[*VertexAttributeArray]*GlVertexAttribPointer,
-	first, count uint32, // vertex indices
-	id api.CmdID,
-	cmd api.Cmd,
-	s *api.State,
-	c *Context,
-	out transform.Writer) {
-
-	if count == 0 {
-		return
-	}
-
-	cb := CommandBuilder{Thread: cmd.Thread()}
-	rngs := interval.U64RangeList{}
-	// Gather together all the client-buffers in use by the vertex-attribs.
-	// Merge together all the memory intervals that these use.
-	va := c.Bound.VertexArray
-	for _, arr := range va.VertexAttributeArrays {
-		if arr.Enabled == GLboolean_GL_TRUE {
-			vb := va.VertexBufferBindings[arr.Binding]
-			if cmd, ok := clientVAs[arr]; ok {
-				// TODO: We're currently ignoring the Offset and Stride fields of the VBB.
-				// TODO: We're currently ignoring the RelativeOffset field of the VA.
-				// TODO: Merge logic with ReadVertexArrays macro in vertex_arrays.api.
-				if vb.Divisor != 0 {
-					panic("Instanced draw calls not currently supported by the compatibility layer")
-				}
-				stride, size := int(cmd.Stride), DataTypeSize(cmd.Type)*int(cmd.Size)
-				if stride == 0 {
-					stride = size
-				}
-				rng := memory.Range{
-					Base: cmd.Data.addr, // Always start from the 0'th vertex to simplify logic.
-					Size: uint64(int(first+count-1)*stride + size),
-				}
-				interval.Merge(&rngs, rng.Span(), true)
-			}
-		}
-	}
-
-	// Create an array-buffer for each chunk of overlapping client-side buffers in
-	// use. These are populated with data below.
-	ids := make([]BufferId, len(rngs))
-	for i := range rngs {
-		ids[i] = t.glGenBuffer(ctx)
-	}
-
-	// Apply the memory observations that were made by the draw call now.
-	// We need to do this as the glBufferData calls below will require the data.
-	dID := id.Derived()
-	out.MutateAndWrite(ctx, dID, cb.Custom(func(ctx context.Context, s *api.State, b *builder.Builder) error {
-		cmd.Extras().Observations().ApplyReads(s.Memory.ApplicationPool())
-		return nil
-	}))
-
-	// Note: be careful of overwriting the observations made above, before the
-	// calls to glBufferData below.
-
-	// Fill the array-buffers with the observed memory data.
-	for i, rng := range rngs {
-		base := memory.BytePtr(rng.First, memory.ApplicationPool)
-		size := GLsizeiptr(rng.Count)
-		t.GlBindBuffer_ArrayBuffer(ctx, ids[i])
-		out.MutateAndWrite(ctx, dID, cb.GlBufferData(GLenum_GL_ARRAY_BUFFER, size, base, GLenum_GL_STATIC_DRAW))
-	}
-
-	// Redirect all the vertex attrib arrays to point to the array-buffer data.
-	for _, l := range va.VertexAttributeArrays.KeysSorted() {
-		arr := va.VertexAttributeArrays[l]
-		if arr.Enabled == GLboolean_GL_TRUE {
-			if cmd, ok := clientVAs[arr]; ok {
-				cmd := *cmd // Copy
-				i := interval.IndexOf(&rngs, cmd.Data.addr)
-				t.GlBindBuffer_ArrayBuffer(ctx, ids[i])
-				cmd.Data = VertexPointer{cmd.Data.addr - rngs[i].First, memory.ApplicationPool} // Offset
-				out.MutateAndWrite(ctx, dID, &cmd)
-			}
-		}
-	}
 }
